move mod: split up mod `arch::mesh_shading` into `mesh` and `task`
