+++ crates/spirv-std/src/arch.rs
@@ -14,7 +14,6 @@ mod atomics;
 mod barrier;
 mod demote_to_helper_invocation_ext;
 mod derivative;
-mod mesh_shading;
 mod ray_tracing;
 mod subgroup;
 
@@ -22,7 +21,6 @@ pub use atomics::*;
 pub use barrier::*;
 pub use demote_to_helper_invocation_ext::*;
 pub use derivative::*;
-pub use mesh_shading::*;
 pub use ray_tracing::*;
 pub use subgroup::*;
